backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "apscheduler>=3.11.2",
    "bcrypt>=5.0.0",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.12.1",
    "pytest>=9.0.2",
    "python-multipart>=0.0.22",
    "sib-api-v3-sdk>=7.6.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
